\name{.Color}
\alias{.Color}
\docType{data}
\title{Colors for Identifying Clusters in Plots}
\description{
  Color themes as used in the package \pkg{lattice}.
}
\usage{.Color}
\format{
  A character vector contains colors used in plots to identify clusters.
}
\references{
  Phylogenetic Clustering Website:
  \url{http://thirteen-01.stat.iastate.edu/snoweye/phyclust/}
}
\author{
  Wei-Chen Chen \email{wccsnow@gmail.com}
}
\seealso{
  \code{\link{plotnj}}.
}
\examples{
\dontrun{
library(phyclust, quiet = TRUE)

.Color
}
}
\keyword{data}
