\name{all.internal}
\alias{bootstrap.merge.seq}
\alias{bootstrap.star.tree}
\alias{bootstrap.seq.nucleotide}
\alias{bootstrap.seq.snp}
\alias{check.EMC}
\alias{convert.QA.to.vect}
\alias{my.cat}
\alias{my.format}
\alias{my.print}
\alias{plotfd}
\alias{plothist.my}
\alias{translate.EMC}
\alias{translate.ret}
\alias{unstrsplit}
\title{ All internal functions of phyclust }
\description{
  See \code{all.internal.tzeng} for Tzeng's methods (2005).

  \code{bootstrap.merge.seq},
  \code{bootstrap.star.tree}, \cr
  \code{bootstrap.seq.nucleotide},
  \code{bootstrap.seq.snp}, \cr
  \code{check.EMC},
  \code{convert.QA.to.vect}, \cr
  \code{my.cat},
  \code{my.format}, \cr
  \code{my.print},
  \code{plotfd}, \cr
  \code{plothist.my},
  \code{translate.EMC}, \cr
  \code{translate.ret},
  \code{unstrsplit}.
}
\references{
  Phylogenetic Clustering Website:
  \url{http://thirteen-01.stat.iastate.edu/snoweye/phyclust/}
}
\author{
  Wei-Chen Chen \email{phyclust@gmail.com}
}
\keyword{internal}
