\name{all.internal.tzeng}
\alias{.tzengEnv}
\alias{chkdim.Add.fun}
\alias{chkdim.Multiply.fun}
\alias{construct.bigB.dPI}
\alias{construct.dB.list.fun}
\alias{final.BigMatB.fun}
\alias{final.BigMatB.matC.fun}
\alias{get.allpossROUTE.fun}
\alias{get.BB.fun}
\alias{get.deriBigMatB.fun}
\alias{get.fullBB.fun}
\alias{get.I.fun}
\alias{get.ini.possHap.hapReserv.fun}
\alias{get.medBB.skip.fun}
\alias{get.preBB.skip.fun}
\alias{get.step.mut.fun}
\alias{get.subPI.fun}
\alias{getBigMatB.fun}
\alias{getI13.fun}
\alias{getPIstar.fun}
\alias{haplo.phase.known}
\alias{I33Inv.fun}
\alias{index.k.fun}
\alias{zero.mat.fun}
\title{ All Internal Functions of Tzeng's Methods }
\description{
  See \code{all.internal} for phyclust.

  \code{chkdim.Add.fun},
  \code{chkdim.Multiply.fun}, \cr
  \code{construct.bigB.dPI},
  \code{construct.dB.list.fun}, \cr
  \code{final.BigMatB.fun},
  \code{final.BigMatB.matC.fun}, \cr
  \code{get.allpossROUTE.fun},
  \code{get.BB.fun}, \cr
  \code{get.deriBigMatB.fun},
  \code{get.fullBB.fun}, \cr
  \code{get.I.fun},
  \code{get.ini.possHap.hapReserv.fun}, \cr
  \code{get.medBB.skip.fun}, \cr
  \code{get.preBB.skip.fun},
  \code{get.step.mut.fun}, \cr
  \code{get.subPI.fun},
  \code{getBigMatB.fun}, \cr
  \code{getI13.fun},
  \code{getPIstar.fun}, \cr
  \code{haplo.phase.known},
  \code{I33Inv.fun},
  \code{index.k.fun}, \cr
  \code{zero.mat.fun}.
}
\references{
  Phylogenetic Clustering Website:
  \url{http://thirteen-01.stat.iastate.edu/snoweye/phyclust/}

  Tzeng, J.Y. (2005)
  \dQuote{Evolutionary-Based Grouping of Haplotypes in Association Analysis},
  \emph{Genetics Epidemiology}, \bold{28}, 220-231.
  \url{http://www4.stat.ncsu.edu/~jytzeng/software.php}
}
\author{
  Jung-Ying Tzeng

  Maintain: Wei-Chen Chen \email{phyclust@gmail.com}
}
\keyword{internal}
\keyword{Haploclust}
