\name{all.internal}
\alias{bootstrap.merge.seq}
\alias{bootstrap.star.tree}
\alias{bootstrap.seq.nucleotide}
\alias{bootstrap.seq.snp}
\alias{check.EMC}
\alias{check.label}
\alias{convert.QA.to.vect}
\alias{my.cat}
\alias{my.format}
\alias{my.print}
\alias{plotfd}
\alias{plothist.my}
\alias{translate.EMC}
\alias{translate.EMC.se}
\alias{translate.ret}
\alias{translate.ret.se}
\alias{unstrsplit}
\alias{code2nid.default}
\alias{code2nid.list}
\alias{nid2code.default}
\alias{nid2code.list}
\alias{code2snp.default}
\alias{code2snp.list}
\alias{snp2code.default}
\alias{snp2code.list}
\alias{sid2snp.default}
\alias{sid2snp.list}
\alias{snp2sid.default}
\alias{snp2sid.list}
\alias{nid2aid.default}
\alias{nid2aid.list}
\alias{nid2cid.default}
\alias{nid2cid.list}
\alias{cid2aid.default}
\alias{cid2aid.list}
\alias{aid2acode.default}
\alias{aid2acode.list}
\alias{acode2aid.default}
\alias{acode2aid.list}
\title{ All Internal Functions of phyclust }
\description{
  See \code{all.internal.tzeng} for Tzeng's methods (2005).

  The following are all internal functions: \cr
  \code{bootstrap.merge.seq},
  \code{bootstrap.star.tree}, \cr
  \code{bootstrap.seq.nucleotide},
  \code{bootstrap.seq.snp}, \cr
  \code{check.EMC},
  \code{check.label},
  \code{convert.QA.to.vect}, \cr
  \code{my.cat},
  \code{my.format}, \cr
  \code{my.print},
  \code{plotfd}, \cr
  \code{plothist.my},
  \code{translate.EMC}, \cr
  \code{translate.ret},
  \code{unstrsplit}.

  The following are all S3 methods: \cr
  \code{code2nid.default},
  \code{code2nid.list}, \cr
  \code{nid2code.default},
  \code{nid2code.list}, \cr
  \code{code2snp.default},
  \code{code2snp.list}, \cr
  \code{snp2code.default},
  \code{snp2code.list}, \cr
  \code{sid2snp.default},
  \code{sid2snp.list}, \cr
  \code{snp2sid.default},
  \code{snp2sid.list}, \cr
  \code{nid2aid.default},
  \code{nid2aid.list}, \cr
  \code{nid2cid.default},
  \code{nid2cid.list}, \cr
  \code{cid2aid.default},
  \code{cid2aid.list}, \cr
  \code{aid2acode.default},
  \code{aid2acode.list}, \cr
  \code{acode2aid.default},
  \code{acode2aid.list}.
}
\references{
  Phylogenetic Clustering Website:
  \url{http://thirteen-01.stat.iastate.edu/snoweye/phyclust/}
}
\author{
  Wei-Chen Chen \email{wccsnow@gmail.com}
}
\keyword{internal}
